requirements.txt
click==8.1.3
rich==13.3.5
GitPython==3.1.31
anthropic==0.3.2
openai==0.27.8
python-dotenv==1.0.0